logic Move uniffi logic and binding generation to the logic itself (#36)
- Having a separate `logic-binding-cpp` module with everything related
to `uniffi` bindings is not scalable as we add more and more to logic
and requires too much duplication. As alternative way we merged binding
generation back to `logic` and rely solely on `uniffi::` proc macros for
binding generation. It creates unfortunately a few issues and complicate
logic a bit, but in a long run it should be much simpler than
maintaining separate UDF file
- One (temporary) issue is that we use `uniffi-bindgen-cpp` which has to
use a specific version of `uniffy` (v0.25) and only starting from
[`uniffy@v.26`](https://github.com/mozilla/uniffi-rs/blob/main/CHANGELOG.md#v0260-backend-crates-v0260---2024-01-23)
we can omit `Arc` from all public interfaces for custom types when proc
macros are used. So for now add a lot of `Arc` which is ugly, but
hopefully temporary and next release of `uniffi-bindgen-cpp` would solve
it
- Also because of `Arc` simple `serde::Serialise/serde::Deserialise`
doesn't work as it doesn't know to work with `Arc<T>` . We anyway wanted
to get rid of JSON serialisation and move to binary encoding, which will
be added next, probably with `bincode`