Robust WebSocket connection (#60)
Took 2 weeks to implement, so let's cut it short and focus only on main
things:

- Robust connection with reconnect and retries
- Even though we are using duplex async communication via websocket with
no sense of Request <> Response the provided API still wraps it in an
incredibly convenient way: Clients sends a message to the backend where
each message has a unique request id. Then backend may send a response
message back to the client with the same request id and websocket class
will call an appropriate callback for that. It makes client side super
simple and easy to reason about:
```c++
//                            [ResponseType      ] [ClientMessage    ]
Connection->SendPublicMessage<logic::ServerStatus>(logic::Ping::init())
//                                [Response is std::variant<ResponseType, ServerError>]
                .Next([](auto Response) {
                    auto Status = std::get<std::shared_ptr<logic::ServerStatus>>(Response);
                    TestEqual("Server status", Status->status(), logic::Status::kOk);
                });
```
- Change server message format - now it has format
`[MessageTag][RequestId][EncodedPayload]`. It allows clients to easily
parse incoming messages and forward it to the right callback
- Clients will assign a unique request id for every outgoing message.
RequestID is u8 counter from 1 to 255, once max is reached it starts
over. We don't expect very high rate of messages, so conflict on
requestId reuse is unlikely to happen
- Use of Automation Spec for testing
https://dev.epicgames.com/documentation/en-us/unreal-engine/automation-spec-in-unreal-engine.
It's much better that `TEST_CASE_NAMED` based we've used in the past and
much more convenient for testing async code
- Rewrite code generation for messages. Previous approach with message
Record + separate Serialiser classes was impossible to work with for
dynamic code generation. Now we have simple message Object with
generated getters for data fields and generated serialise/deserialise
code